Implement variable length integer support
This patch provides an implementation of variable-length integers,
both in libdtrace and as precompiled BPF code.  It is based on the
PrefixVarint algorithm.  Unsigned integers are encoded as follows:

  0x0 - 0x7f		1 byte:  0xxxxxxx
  0x80 - 0x407f		2 bytes: 10xxxxxx xxxxxxxx
  0x4080 - 0x20407f	3 bytes: 110xxxxx xxxxxxxx xxxxxxxx

and so on.  This algorithm ensures that there is exactly one
representation for a given unsigned integer.  It also ensures that
the number of bytes in the variable length representation can be
determined based on the first byte.  This makes decoding simpler.

The following functions are provided:

  int dt_int2vint(uint64_t val, char *str)
	Store a 64-bit unsigned integer in the provided string and
	return the number of bytes used.

  uint64_t dt_vint2int(const char *str)
	Retrieve a 64-bit unsigned integer from a variable length
	integer in the provided string.

  int dt_vint_size(uint64_t val)
	Return the number of bytes necessary to represent the given
	64-bit unsigned integer as a variable length integer.

  const char *dt_vint_skip(const char *str)
	Skip over the variable length integer at the start of the
	given string and return a pointer to the first byte beyond
	the variable length integer.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>