Store strings in the string table with varint length prefix
With the introduction of variable length integers, strings in D (and
then also BPF) can be stored with their length embedded in the byte
stream.  This makes it possible to perform operations on strings
without the need to recalculate the length of the string multiple
times.

The string constant table stored in DIFOs contains strings that will
be made available to the BPF program by loading the string constant
table into a BPF map.  It makes sense to already store strings using
the variable length prefix when the string constant table is
constructed, which then also requires code that uses it to know
about the variable length integer prefixing each string.

This patch introduces a dt_difo_getstr() function that returns a
regular C char * entity given a DIFO and a string offset into the
string constant table for that DIFO.  Any code that needs to access
a string constant from a DIFO is updated to make use of this new
function.

The string constant table creation code is updated to account for
the fact that the empty string will now occupy 2 bytes (one for the
length that is 0, and one for the terminating 0-byte.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>