Support probeprov, probemod, probefunc, probename built-in variables
This patch adds support for the built-in variables that provide the
component names for the probe description of the current probe in a
D clause, i.e. the probe that triggered the execution of the clause.

Because clauses (compiled into BPF functions) can be linked into the
BPF programs for multiple probes, the value of these built-in variables
cannot be provided at compile time.  Their value is obtained from a
new BPF hashmap named 'probes' that associates a probe id with the
constant strings that provide the components of the probe description.

This patch also introduces a new BPF constant (STBSZ, resolved at link
time) to hold the size of the string constant table.  This value is
needed to satisfy BPF verifier expectations for map value access
boundary checking.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>