Aggregations are not data-recording actions
A clause like { @ = count() } should not be data-recording.  The
whole point of aggregations is that we should not have to see output
for every probe firing, even if -q/quiet is turned off.  Nevertheless,
we still have to vet clauses to make sure that, e.g., a clause does
not combine aggregations and speculation.

The reason this problem was not caught earlier is that testing is
typically done in "quiet" mode.

Add a new DT_CLSFLAG_AGGREGATION and split this case off of
DT_CLSFLAG_DATAREC in clause-flag processing.  Add a test for
aggregations that does not use quiet mode.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>