Do not print an offset for builtin variables
Builtin variables are global variables but their value is not stored
in the gvars BPF map.  Therefore, their offset is always -1 which is
displayed as ffffffff.  That value is too wide for the OFFSET field
in the variable list in the disassembly listing.  Since the offset
value is not meaningful in this case, it is no longer printed.

Offsets are now also printed in decimal rather than hexadecimal to be
consistent with other offset values in disassembly listings.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>