Correct storage size assignment for variables
The implementation of global and local variables using BPF maps added
calls to dt_ident_set_storage() to set the offset and size of the data
storage for each variable.  It was setting the storage requirement for
any automatically declared variable to 8 bytes regardless of its type.
Since the datatype of such a variable is determined dynamically based
on its use, the storage assignment must be done once the datatype is
known.  But by then, some code was already generated based on the wrong
size.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>