Add support for the ustack() action
Stack unwind is very difficult.  Therefore, adapt testing to the safest of
conditions:  probe firing in a stack that we compiled with no optimization,
-fno-omit-frame-pointer, no tail call optimizations, and no leaf function
optimizations.  Further, adapt the test to take better advantage of test
suite infrastructure for triggers and results checking.

The "strsize" argument has no effect.  Its value is silently ignored, and
strsize==0 is assumed.

When a pid entry probe fires, the stack frame is not yet set up.  Therefore,
the immediate caller will be missing from the call stack.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>