Optimize the memcpy implementation and make it more robust
The memcpy() implementation in BPF was more complicated than it needed
to be.  As a result, the BPF verifier was not able to always guarantee
that this function was being used in a safe way.  The simplification of
the implementation resolves this.

This patch also enforces a maximum copy length of 256 bytes,  In the
future, this limit surely needs to be increased.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>