Framework for D subroutines incl. strlen() and pseudo speculation()
This patch provides the skeleton for the implementation of subroutines
in D.  It is equivalent to the implementation of D actions in the code
generator.

The strlen() subroutine is provided as a first example of how to
imeplement subroutines.  The speculation() subroutine is provided as
an example of a dummy implementation.  It always simply returns 0, but
this allows various tests for the compiler to be exercised.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>