Increase the instruction counter to 4 digits for BPF programs
BPF programs are becoming longer as more features are implemented, and
we are now routinely encountering progrmas that are 1000 or more
instructions.  The disassembler listings were limited to 3 digits for
the instruction counter and 4 for the offset.  This is now increased
to 4 digits for the counter and 5 for the offset.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>