Add support for built-in variable walltimestamp
In the pre-compiled function get_bvar(), use bpf_ktime_get_ns() to
get the time (in nsec) since boot.  This value is adjusted by adding
the POSIX time when the system was booted.

A few more walltimestamp tests pass now, but existing walltimestamp
tests are rather lenient.  Add a new, more stringent test.  In particular,
timestamp is supposed to be fixed for an entire clause while walltimestamp
is fresh with every call.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>