Implement argument retrieval for SDT probes
The arguments for tracepoints are presented in the BPF context as a
sequence of 64-bit values.  They are preceded by a double word (64-bit
value) that is not accessible from BPF programs.

The tst.args.d test case verifies that we can retrieve argument data
for a tracepoint.  It is quite complex because we end up having to
decode the argument data because of the component datatypes.  This
adds to the value of the test because it verifies that the argument
data is presented in the correct way.

Both tst.arg-arity.d and tst.args-sdt.d are removed because they are
specific to the previous implementation of DTrace on Linux, and depend
on a deprecated kernel-level probe that existed for the sole purpose of
being able to use these tetss.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>