Provide set_upper_bound() and set_lower_bound() macro for C-to-BPF
The BPF verifier obtains (and needs) information about upper and lower
bounds for values based on conditional clauses.  The BPF GCC cross
compiler generates conditionals in a variety of forms.  The BPF verifier
sypports only a limited set of conditionals, so we need to be able to
enforce a specific form.  More specifically, the BPF verifier requires
the bound (a constant value) to be stored in the src register while the
value to be validated is in the dst register.

This patch provides macros that expand to inline assembler instructions
that enforce the expected register convention.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>