consume, cg: implement speculations
This works by tracking live speculations in a specs map containing
dt_bpf_specs_t, each of which tracks the number of buffers written and
whether a commit/discard has been called on it.  (bpf/speculation.c,
which finds free speculations, is limited to 16 speculations and needs
adjusting when BPF loops work).  Successful speculate()s record the ID
of the active speculation in the clause header.  When a commit happens,
a COMMIT/DISCARD record is written.

At consume time, buffers with active speculate()s are hived off into a
list of dt_spec_bufs_t under a dt_spec_bufs_head_t (when peeking, this
is only done on the first peek, identified via the CONSUME_PEEK_START
peekflag); when a commit/discard hits for a given speculation, the efunc
invocation etc is suppressed if need be (discards can be followed by
data-recording actions, in which case the efunc is not suppressed), and
the specs map entry for this speculation is sucked into the
dt_spec_bufs_head and the head chained into the dt_spec_bufs_draining
list.  This list is traversed on every consume, and any new spec bufs
committed (as if just received from the ring buffer) or discarded
appropriately.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>