tests: reduce timeout of very slow test
This test sets the timeout to a high value to fill a buffer that no
longer applies to v2, and is marked @@xfail: dtv2.  Drop the timeout so
we don't have to wait ten minutes on every test run for something we
know will fail.

We can raise it again later if it seems necessary.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>