Fix data size for value copy in dt_cg_store_var()
The size of the data to be copied in dt_cg_store_var() was determined
wrongly.  For strings constants we were copying the size of the string
constant even if that might be larger than the maximum string size.

The offsets for lvars and gvars was also not taking into account the
length prefix to be stored for each string value.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>