htab reduction: kernpath
Replace the special-purpose kernpath hash table with the generic
dt_htab.  This is a simple case which serves as a template for the
more complex ones to follow.

This has very little effect on performance, because even though the htab
is terribly poorly sized, it is sized in the wrong direction: the
kernpath hash only contains out-of-tree modules, and there won't be
hundreds of those.  The main improvement here is in clarity.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>