cmd: delete ctf_module_dump
This tool was useful back in the day because it knew how to read CTF
that was dumped into kernel modules, digging out the shared CTF
dict and built-in module info from the fake ctf.ko module and
passing it to ctf_dump.

But it is 2021.  We haven't linked CTF into in-tree kernel modules since
UEK4 in 2017 (4.1.12-113, libdtrace-ctf 0.7).  Every tool needed for
this is obsolete, including libdtrace-ctf's ctf_dump; and objdump does
just as good a job of dumping modules as this tool (or, actually, given
that it can't dump raw .ctfa files yet, just as bad a job; but this will
be fixed soon and is easily workable around with a single objcopy call
even now).  Also ctf_module_dump is getting in the way of
representational improvements to the kernel module tracking.

So drop ctf_module_dump.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Suggested-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>