Ensure proper validation checks are done before freeing a tstring
We want to call dt_cg_tstring_free() if a tstring is in use.  Two
styles of checks were being used:
    if (dnp->dn_kind == DT_NODE_FUNC && dnp->dn_tstring)
        dt_cg_tstring_free(pcb, dnp);
and
    if (dnp->dn_tstring)
        dt_cg_tstring_free(pcb, dnp);
The problem with the second style is that if dn_kind is not DT_NODE_FUNC,
then dnp->dn_string might still be non-NULL since it is a union with
other members.  This leads to calling dt_cg_tstring_free() with an
invalid address and an assertion failure.

Add the check for dnp->dn_kind == DT_NODE_FUNC and dnp->dn_tstring into
dt_cg_tstring_free() to ensure the checks are performed at all times and
to simplify the code.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>