Tweak testsuite to account for libctf / libdtrace-ctf differences
Some error messages changed in libctf while libdrace-ctf is still using
the old messages.  The most common change is the addition of a trailing
period for error messages.

Forward declarations were printed as 'struct FOO' with libdtrace-ctf
even if they were for a union or enum.  With libctf they are printed
correctly - the testsuite tests affected by this now support both cases.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>