Adjust dynvarsize to avoid hitting BPF size limits
The number of entries in the dvars BPF map is calculated as the
dynvarsize divided by the size of the value type, which is 1.  It
turns out that there is a limit on the number of entries and this
test hits that limit.

This patch sets dynvarsize to 1024 to force a lower limit so that
this issue is not triggered anymore.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>