Avoid checking for NULL pointers multiple times
The dt_cg_load_var() function included a check to see that a loaded
pointer value was not NULL (and to generate a probe error if it was).
However, callers of this function are already checking for NULL before
using the pointer (if such a check is necessary).  Therefore, this
does not need to be done in dt_cg_load_var().

The dt_cg_load_var() function was also creating two labels that were
only actually used in one branch.  Eugene Loh also pointed out that a
MOV instruction could be saved here.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>