test: use more reliable way of finding kernel versions
In several places we were using cut and/or awk to find kernel versions:
in each case the code used worked for versions like 5.15 or 5.11.4 but
failed for things like 5.15+ or 5.15-00024-g679b45e057b4-dirty,
because they are splitting fields on dots alone.

The rarely-used and little-known grep -o can do all of this more simply
and reliably, since it can split fields using full regular expressions
(and thus can easily be told that what we are looking for here is a run
of digits only).

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>