Remove the string length prefix
This is the final stage of discontinuing the string length prefix code.
It removes the actual storage allocation for the 2-byte string length
prefix, and updates all code that adjusted pointers because of the
prefix.  The dt_strlen_store() function is gone.

This change affects the offset calculation for variab;e storage when a
string is involved.  Test results for gvar/tst.undecl-offset.r are
updated to reflect this.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>