Pass load-time constants to BPF programs in a consistent way
In bpf/, there are .c and .S programs that are cross-compiled at
build time, but they sometimes rely on constants that are not known
until the programs are loaded.  Two ways were being used to provide
those constants.  One was to put them in the state map.  Another was
to set the values during relocation.

Use one consistent method for providing these constants.

Specifically, set the constants during relocation rather than using
the state map:

*)  These constants do not change during execution.

*)  Avoiding a map lookup saves a function call, saving BPF
    instructions, simplifying BPF programming, and reducing
    register pressure.

*)  The constants' values are made known to the BPF verifier.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>