test: Fix error tag in test name
A number of err.D_*.d tests had the wrong error tag in the test
name, which escaped detection for a long time because runtest.sh
did not check the error tags.  A number of such cases were fixed.
This appears to be the last case.  The test encounters D_SYNTAX,
both in legacy DTrace on Linux and in the current version.  Rename
the test appropriately.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>