Add support for mutex-related subroutines
Implement the relevant subroutines.  Clean up the tests.

A few tests were broken in ways that were not being detected:
early on because runtest.sh was not checking error tags and later
because the tests were marked XFAIL.  Clean these tests up.

Changes include not using kmutex_t any longer in the tests.  This
type was not recognized even in legacy DTrace.  The documentation
needs a corresponding update.

Similarly, probe lockstat:genunix:mutex_enter:adaptive-acquire was
not recognized, even by legacy DTrace, and it was not needed for the
tests in question.

There were no rw_* tests.  One has been added, but it remains
unclear how stable the test is -- specifically, whether enough
write-lock events are generated.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>