Fix stack offsets for clause functions
Clause functions were reserving space on the stack for the DTrace
context even though they only store a pointer to the DTrace context
that is stored on the stack of the trampoline program.  This wasted
64 bytes (out of 512 bytes that are available).

This patch also provides some reworking of how reference is made to
members of dctx on the stack.  Both trampolines and functions now use
[reg + offset].

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>