bpf: don't throw away CTF errors when doing relocations
The first type lookup done after translators are parsed is the set done
to relocate CTF-related offsets.  If something is wrong with the CTF and
the translators are missing, these lookups are likely to fail -- but
since most of these lookups don't set dt_ctferr, the actual error is
discarded and the only thing reported to the user is "unknown CTF
error".  Not even CTF debugging helps.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>